mini_projects/jozsef.gulyas/homework.01/pyproject.toml
[project]
name = "homework-01"
version = "0.1.0"
description = "First homework. Console application. Call an api (openai in this case), and return with a hr response."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
